% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readStaticInterfileIMG.R
\name{readStaticInterfileIMG}
\alias{readStaticInterfileIMG}
\title{Read interfile binary file format images}
\usage{
readStaticInterfileIMG(fname, endian = "little", file.row = 80,
  file.col = 80, what = "double", bytes.per.pixel = 4)
}
\arguments{
\item{fname}{Filename of the interfile}

\item{endian}{Endian-ness of the file you are attempting to read. Permitted values are "big" or "little"}

\item{file.row}{Rows of pixels in the image}

\item{file.col}{Columns of pixels in the image}

\item{what}{Type of binary data stored: permitted values are "numeric", "double", "integer", "int", "logical", "complex", "character", "raw".}

\item{bytes.per.pixel}{Number of bytes per pixel. This information shold be stored in the header file (.hdr) of the interfile.}
}
\value{
A matrix of pixel intensities
}
\description{
This function imports static (2D) binary interfile files.
}
\examples{

# Load an example interfile (Rabbit aorta, Tc99m scan)
interfile.location <- system.file("extdata", "Tc99m_20percent_energy.img", package = "interfile")

# Read the binary interfile
my_scan <- readStaticInterfileIMG(interfile.location)

# Check out the results (returned as a matrix)
dim(my_scan)
head(my_scan)
}
